pyproject.toml
[project]
requires-python = "==3.13.*"
name = "bot"
description = "Deletes messages."
version = "1.0.0"
authors = [{ name = "BlindfoldedSurgery", email = "team@blindfolded.surgery" }]
classifiers = [
    "Private :: Do Not Upload",
    "License :: OSI Approved :: MIT License",
]

dependencies = [
    "bs-config [dotenv] ==3.2.1",
    "bs-nats-updater ==3.0.0",
    "bs-state [redis] ==2.4.*",
    "httpx ==0.28.*",
    "pydantic ==2.12.*",
    "python-telegram-bot ==22.5",
    "pyyaml ==6.0.3",
    "replicate==1.0.7",
    "sentry-sdk >=2.0.0, <3.0.0",
    "uvloop ==0.22.*",
]

[dependency-groups]
dev = [
    "mypy ==1.18.*",
    "pytest >=8.0.0, <9.0.0",
    "ruff ==0.14.2",
    "types-pyyaml >=6, <7",
]

[[tool.uv.index]]
name = "internal"
url = "https://pypi.bjoernpetersen.net/simple"
explicit = true

[tool.uv.sources]
bs-config = { index = "internal" }
bs-nats-updater = { index = "internal" }
bs-state = { index = "internal" }

[build-system]
requires = ["uv_build==0.9.*"]
build-backend = "uv_build"

[tool.mypy]
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "google.cloud.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "scipy"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "cv2"
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true

[tool.ruff.lint]
select = [
    "E4",
    "E7",
    "E9",
    "F",
    "FURB",
    "G",
    "I001",
    "LOG",
    "PTH",
    "Q",
    "UP",
    "TC001",
    "TC002",
    "TC003",
    "TC004",
    "TC005",
]
ignore = [
    # redundant-open-modes
    "UP015",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
